+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-# CHANGES
+# Changelog
 
-## v4.4.0-RC
+## [Unreleased](https://github.com/weseek/growi/compare/v4.3.3...HEAD)
+
+*Please do not manually update this file. We've automated the process.*
 
 ### BREAKING CHANGES
 
@@ -8,6 +10,8 @@
 
 ### Updates
 
+* Feature: Password resetting by user
+* Feature: User trigger notification and Global notification are available by new Slack integration
 * Improvement: Add attachment button in editor navbar
 * Fix: Recursive rename operation from `/parent` to `/parent/child` ([#4101](https://github.com/weseek/growi/pull/4101))
 * Fix: Encode spaces in page path in LinkEditModal
@@ -22,7 +26,7 @@
     * morgan
     * socket.io
 
-## v4.3.3-RC
+## v4.3.3
 
 * Improvement: Welcome page markdown
 * Fix: Some recursive operation exclude descendant pages that are restricted for groups